Move queries for sponsors list command from api package to list package
This is to conform with the existing patterns in the repo
